refactor: decouple pipeline from actions to resolve require cycles
Introduce DeltaActionPort in lib/pipeline to break value-level circular
imports between the pipeline and actions modules. This resolves Metro
require cycles that lead to Out-Of-Memory compilation loops during
bundling. Eager setup module graphs inside vitest.setup.ts are also
decoupled using concrete sub-file imports to keep mock contexts isolated.
