+++ lib/db/index.ts
@@ -147,5 +147,5 @@ export type {
   VaultCalendar,
 } from './types'
 export { runInTransaction } from './runtime/transaction'
-export type { SqlOp } from './types'
+export type { DbCtx, SqlOp } from './types'
 export { useDbMigrations } from './runtime/use-db-migrations'
