+++ lib/pipeline/runtime/recovery.ts
@@ -1,7 +1,6 @@
 import { asc, eq, isNull } from 'drizzle-orm'
 
-import type { DbCtx } from '@/lib/actions'
-import { pipelineRuns } from '@/lib/db'
+import { pipelineRuns, type DbCtx } from '@/lib/db'
 import { logger } from '@/lib/diagnostics'
 
 import { reverseReplayDeltas } from './action-port'
